Struct ext_php_rs::bindings::_zend_class_entry

#[repr(C)]
pub struct _zend_class_entry {

    pub type_: c_char,
    pub name: *mut zend_string,
    pub __bindgen_anon_1: _zend_class_entry__bindgen_ty_1,
    pub refcount: c_int,
    pub ce_flags: u32,
    pub default_properties_count: c_int,
    pub default_static_members_count: c_int,
    pub default_properties_table: *mut zval,
    pub default_static_members_table: *mut zval,
    pub static_members_table__ptr: *mut *mut zval,
    pub function_table: HashTable,
    pub properties_info: HashTable,
    pub constants_table: HashTable,
    pub properties_info_table: *mut *mut _zend_property_info,
    pub constructor: *mut zend_function,
    pub destructor: *mut zend_function,
    pub clone: *mut zend_function,
    pub __get: *mut zend_function,
    pub __set: *mut zend_function,
    pub __unset: *mut zend_function,
    pub __isset: *mut zend_function,
    pub __call: *mut zend_function,
    pub __callstatic: *mut zend_function,
    pub __tostring: *mut zend_function,
    pub __debugInfo: *mut zend_function,
    pub __serialize: *mut zend_function,
    pub __unserialize: *mut zend_function,
    pub iterator_funcs_ptr: *mut zend_class_iterator_funcs,
    pub __bindgen_anon_2: _zend_class_entry__bindgen_ty_2,
    pub get_iterator: Option<unsafe extern "C" fn(ce: *mut zend_class_entry, object: *mut zval, by_ref: c_int) -> *mut zend_object_iterator>,
    pub get_static_method: Option<unsafe extern "C" fn(ce: *mut zend_class_entry, method: *mut zend_string) -> *mut zend_function>,
    pub serialize: Option<unsafe extern "C" fn(object: *mut zval, buffer: *mut *mut c_uchar, buf_len: *mut size_t, data: *mut zend_serialize_data) -> c_int>,
    pub unserialize: Option<unsafe extern "C" fn(object: *mut zval, ce: *mut zend_class_entry, buf: *const c_uchar, buf_len: size_t, data: *mut zend_unserialize_data) -> c_int>,
    pub num_interfaces: u32,
    pub num_traits: u32,
    pub __bindgen_anon_3: _zend_class_entry__bindgen_ty_3,
    pub trait_names: *mut zend_class_name,
    pub trait_aliases: *mut *mut zend_trait_alias,
    pub trait_precedences: *mut *mut zend_trait_precedence,
    pub attributes: *mut HashTable,
    pub info: _zend_class_entry__bindgen_ty_4,

}

Fields
type_: c_charname: *mut zend_string__bindgen_anon_1: _zend_class_entry__bindgen_ty_1refcount: c_intce_flags: u32default_properties_count: c_intdefault_static_members_count: c_intdefault_properties_table: *mut zvaldefault_static_members_table: *mut zvalstatic_members_table__ptr: *mut *mut zvalfunction_table: HashTableproperties_info: HashTableconstants_table: HashTableproperties_info_table: *mut *mut _zend_property_infoconstructor: *mut zend_functiondestructor: *mut zend_functionclone: *mut zend_function__get: *mut zend_function__set: *mut zend_function__unset: *mut zend_function__isset: *mut zend_function__call: *mut zend_function__callstatic: *mut zend_function__tostring: *mut zend_function__debugInfo: *mut zend_function__serialize: *mut zend_function__unserialize: *mut zend_functioniterator_funcs_ptr: *mut zend_class_iterator_funcs__bindgen_anon_2: _zend_class_entry__bindgen_ty_2get_iterator: Option<unsafe extern "C" fn(ce: *mut zend_class_entry, object: *mut zval, by_ref: c_int) -> *mut zend_object_iterator>get_static_method: Option<unsafe extern "C" fn(ce: *mut zend_class_entry, method: *mut zend_string) -> *mut zend_function>serialize: Option<unsafe extern "C" fn(object: *mut zval, buffer: *mut *mut c_uchar, buf_len: *mut size_t, data: *mut zend_serialize_data) -> c_int>unserialize: Option<unsafe extern "C" fn(object: *mut zval, ce: *mut zend_class_entry, buf: *const c_uchar, buf_len: size_t, data: *mut zend_unserialize_data) -> c_int>num_interfaces: u32num_traits: u32__bindgen_anon_3: _zend_class_entry__bindgen_ty_3trait_names: *mut zend_class_nametrait_aliases: *mut *mut zend_trait_aliastrait_precedences: *mut *mut zend_trait_precedenceattributes: *mut HashTableinfo: _zend_class_entry__bindgen_ty_4
Implementations

impl _zend_class_entry

pub fn try_find(name: &str) -> Option<&'static Self>

Attempts to find a reference to a class in the global class table.

Returns a reference to the class if found, or None if the class could not be found or the class table has not been initialized.

pub fn flags(&self) -> ClassFlags

Returns the class flags.

pub fn is_interface(&self) -> bool

Returns true if the class entry is an interface, and false otherwise.

pub fn instance_of(&self, ce: &ClassEntry) -> bool

Checks if the class is an instance of another class or interface.

Parameters

ce - The inherited class entry to check.

pub fn interfaces(&self) -> Option<impl Iterator<Item = &ClassEntry>>

Returns an iterator of all the interfaces that the class implements. Returns None if the interfaces have not been resolved on the class.

pub fn parent(&self) -> Option<&Self>

Returns the parent of the class.

If the parent of the class has not been resolved, it attempts to find the parent by name. Returns None if the parent was not resolved and the parent was not able to be found by name.

impl _zend_class_entry

pub fn throwable() -> &'static Self

Returns the base Throwable class.

pub fn exception() -> &'static Self

Returns the base Exception class.

pub fn error_exception() -> &'static Self

Returns the base ErrorException class.

pub fn compile_error() -> &'static Self

Returns the base CompileError class.

pub fn parse_error() -> &'static Self

Returns the base ParseError class.

pub fn type_error() -> &'static Self

Returns the base TypeError class.

pub fn argument_count_error() -> &'static Self

Returns the base ArgumentCountError class.

pub fn value_error() -> &'static Self

Returns the base ValueError class.

pub fn arithmetic_error() -> &'static Self

Returns the base ArithmeticError class.

pub fn division_by_zero_error() -> &'static Self

Returns the base DivisionByZeroError class.

pub fn unhandled_match_error() -> &'static Self

Returns the base UnhandledMatchError class.