Struct ext_php_rs::bindings::_zend_execute_data

#[repr(C)]
pub struct _zend_execute_data {
    pub opline: *const zend_op,
    pub call: *mut zend_execute_data,
    pub return_value: *mut zval,
    pub func: *mut zend_function,
    pub This: zval,
    pub prev_execute_data: *mut zend_execute_data,
    pub symbol_table: *mut zend_array,
    pub run_time_cache: *mut *mut c_void,
    pub extra_named_params: *mut zend_array,
}

Fields
opline: *const zend_opcall: *mut zend_execute_datareturn_value: *mut zvalfunc: *mut zend_functionThis: zvalprev_execute_data: *mut zend_execute_datasymbol_table: *mut zend_arrayrun_time_cache: *mut *mut c_voidextra_named_params: *mut zend_array
Implementations

impl _zend_execute_data

pub unsafe fn get_object<T: RegisteredClass>(
&self
) -> Option<ClassObject<'static, T>>

Attempts to retrieve a reference to the underlying class object of the Zend object.

Returns a ClassObject if the execution data contained a valid object, otherwise returns None.

Safety

The caller must guarantee that the function is called on an instance of ExecutionData that:

Contains an object.

The object was originally derived from T.

pub fn get_self(&self) -> Result<&mut ZendObject>

Attempts to retrieve the ‘this’ object, which can be used in class methods to retrieve the underlying Zend object.