Struct ext_php_rs::bindings::_zend_object_handlers

#[repr(C)]pub struct _zend_object_handlers {
    pub offset: c_int,
    pub free_obj: zend_object_free_obj_t,
    pub dtor_obj: zend_object_dtor_obj_t,
    pub clone_obj: zend_object_clone_obj_t,
    pub read_property: zend_object_read_property_t,
    pub write_property: zend_object_write_property_t,
    pub read_dimension: zend_object_read_dimension_t,
    pub write_dimension: zend_object_write_dimension_t,
    pub get_property_ptr_ptr: zend_object_get_property_ptr_ptr_t,
    pub has_property: zend_object_has_property_t,
    pub unset_property: zend_object_unset_property_t,
    pub has_dimension: zend_object_has_dimension_t,
    pub unset_dimension: zend_object_unset_dimension_t,
    pub get_properties: zend_object_get_properties_t,
    pub get_method: zend_object_get_method_t,
    pub get_constructor: zend_object_get_constructor_t,
    pub get_class_name: zend_object_get_class_name_t,
    pub cast_object: zend_object_cast_t,
    pub count_elements: zend_object_count_elements_t,
    pub get_debug_info: zend_object_get_debug_info_t,
    pub get_closure: zend_object_get_closure_t,
    pub get_gc: zend_object_get_gc_t,
    pub do_operation: zend_object_do_operation_t,
    pub compare: zend_object_compare_t,
    pub get_properties_for: zend_object_get_properties_for_t,
}

Fields
offset: c_intfree_obj: zend_object_free_obj_tdtor_obj: zend_object_dtor_obj_tclone_obj: zend_object_clone_obj_tread_property: zend_object_read_property_twrite_property: zend_object_write_property_tread_dimension: zend_object_read_dimension_twrite_dimension: zend_object_write_dimension_tget_property_ptr_ptr: zend_object_get_property_ptr_ptr_thas_property: zend_object_has_property_tunset_property: zend_object_unset_property_thas_dimension: zend_object_has_dimension_tunset_dimension: zend_object_unset_dimension_tget_properties: zend_object_get_properties_tget_method: zend_object_get_method_tget_constructor: zend_object_get_constructor_tget_class_name: zend_object_get_class_name_tcast_object: zend_object_cast_tcount_elements: zend_object_count_elements_tget_debug_info: zend_object_get_debug_info_tget_closure: zend_object_get_closure_tget_gc: zend_object_get_gc_tdo_operation: zend_object_do_operation_tcompare: zend_object_compare_tget_properties_for: zend_object_get_properties_for_t
Implementations

impl _zend_object_handlers

pub fn init<T>() -> *mut ZendObjectHandlers

Creates a new set of object handlers from the standard object handlers, returning a pointer to the handlers.