Module ext_php_rs::php

Objects relating to PHP and the Zend engine.

Modules

args
Builder and objects relating to function and method arguments.

class
Builder and objects for creating classes in the PHP world.

constants

enums
Wrapper for enums introduced in C.

exceptions
Contains all the base PHP throwables, including Throwable and Exception.

execution_data
Functions for interacting with the execution data passed to PHP functions
introduced in Rust.

flags
Bitflags used in PHP and the Zend engine.

function
Builder and objects used to create functions and methods in PHP.

module
Builder and objects for creating modules in PHP. A module is the base of a PHP extension.

pack
Provides implementations for converting to and from Zend binary strings, commonly returned from functions such as pack and unpack.

types
Contains all the different types that are introduced into PHP. Introduces functions for converting between Zend values and Rust values.