Struct ext_php_rs::php::class::ClassBuilder

pub struct ClassBuilder<'a> { /* fields omitted */ }

Builds a class to be exported as a PHP class.

Implementations

impl<'a> ClassBuilder<'a>

pub fn new<N>(name: N) -> Self where
N: AsRef<str>,

Creates a new class builder, used to build classes to be exported to PHP.

Parameters

name - The name of the class.

pub fn extends(self, parent: *mut ClassEntry) -> Self

Sets the class builder to extend another class.

Parameters

parent - The parent class to extend.

pub fn method(self, func: FunctionEntry, flags: MethodFlags) -> Self

Adds a method to the class.

Parameters

func - The function entry to add to the class.

flags - Flags relating to the function. See MethodFlags.

pub fn property(
self,
name: impl AsRef<str>,
default: impl Into<Zval>,
flags: PropertyFlags
) -> Self

Adds a property to the class. The initial type of the property is given by the type of the given default. Note that the user can change the type.

Parameters

name - The name of the property to add to the class.

default - The default value of the property.

flags - Flags relating to the property. See PropertyFlags.

pub fn constant(self, name: impl AsRef<str>, value: impl Into<Zval>) -> Self

Adds a constant to the class. The type of the constant is defined by the type of the given default.

Parameters

name - The name of the constant to add to the class.

value - The value of the constant.

pub fn flags(self, flags: ClassFlags) -> Self

Sets the flags for the class.

Parameters

flags - Flags relating to the class. See ClassFlags.

pub fn object_override<T>(self) -> Self where
T: ZendObjectOverride,

Overrides the creation of the Zend object which will represent an instance of this class.

Parameters

T - The type which will override the Zend object. Must implement ZendObjectOverride which can be derived through the ZendObjectHandler derive macro.

pub fn build(self) -> *mut ClassEntry

Builds the class, returning a pointer to the class entry.