Type Definition ext_php_rs::php::class::ClassEntry

type ClassEntry = zend_class_entry;

A Zend class entry. Alias.

Implementations

impl ClassEntry

pub fn try_find(name: &str) -> Option<&'static Self>

Attempts to find a reference to a class in the global class table.

Returns a reference to the class if found, or None if the class could not be found or the class table has not been initialized.

pub fn flags(&self) -> ClassFlags

Returns the class flags.

pub fn instance_of(&self, ce: &ClassEntry) -> bool

Checks if the class is an instance of another class or interface.

Parameters

ce - The inherited class entry to check.

pub fn interfaces(&self) -> Option<impl Iterator<Item = &ClassEntry>>

Returns an iterator of all the interfaces that the class implements. Returns None if the interfaces have not been resolved on the class.

pub fn parent(&self) -> Option<&Self>

Returns the parent of the class.

If the parent of the class has not been resolved, it attempts to find the parent by name. Returns None if the parent was not resolved and the parent was not able to be found by name.

impl ClassEntry

pub fn throwable() -> &'static Self

Returns the base Throwable class.

pub fn exception() -> &'static Self

Returns the base Exception class.

pub fn error_exception() -> &'static Self

Returns the base ErrorException class.

pub fn compile_error() -> &'static Self

Returns the base CompileError class.

pub fn parse_error() -> &'static Self

Returns the base ParseError class.

pub fn type_error() -> &'static Self

Returns the base TypeError class.

pub fn argument_count_error() -> &'static Self

Returns the base ArgumentCountError class.

pub fn value_error() -> &'static Self

Returns the base ValueError class.

pub fn arithmetic_error() -> &'static Self

Returns the base ArithmeticError class.

pub fn division_by_zero_error() -> &'static Self

Returns the base DivisionByZeroError class.

pub fn unhandled_match_error() -> &'static Self

Returns the base UnhandledMatchError class.