Type Definition ext_php_rs::php::execution_data::ExecutionData

type ExecutionData = zend_execute_data;

Execution data passed when a function is called from Zend.

Implementations

impl ExecutionData

pub unsafe fn get_object<T: RegisteredClass>(
&self
) -> Option<ClassObject<'static, T>>

Attempts to retrieve a reference to the underlying class object of the Zend object.

Returns a ClassObject if the execution data contained a valid object, otherwise returns None.

Safety

The caller must guarantee that the function is called on an instance of ExecutionData that:

Contains an object.

The object was originally derived from T.

pub fn get_self(&self) -> Result<&mut ZendObject>

Attempts to retrieve the ‘this’ object, which can be used in class methods to retrieve the underlying Zend object.