Type Definition ext_php_rs::php::function::FunctionEntry

type FunctionEntry = zend_function_entry;

A Zend function entry. Alias.

Implementations

impl FunctionEntry

pub fn end() -> Self

Returns an empty function entry, signifing the end of a function list.

pub fn into_raw(self) -> *mut Self

Converts the function entry into a raw and pointer, releasing it to the C world.