Module ext_php_rs::php::types

Contains all the different types that are introduced into PHP. Introduces functions for converting between Zend values and Rust values.

Modules

array
Represents an array in PHP. As all arrays in PHP are associative arrays, they are represented by hash tables.

long
Represents an integer introduced in PHP. Note that the size of this integer differs. On a 32-bit system, a ZendLong is 32-bits, while on a 64-bit system it is 64-bits.

object
Represents an object in PHP. Allows for overriding the internal object used by classes, allowing users to store Rust data inside a PHP object.

string
Represents a string in the PHP world. Similar to a C string, but is reference counted and contains the length of the string, meaning the string can contain the NUL character.

zval
The base value in PHP. A Zval can contain any PHP type, and the type that it contains is determined by a property inside the struct. The content of the Zval is stored in a union.

Type Definitions

ZendType
Internal Zend type.