Struct ext_php_rs::php::types::array::ZendHashTable

pub struct ZendHashTable { /* fields omitted */ }

A PHP array, which internally is a hash table.

Implementations

impl ZendHashTable

pub fn new() -> Self

Creates a new, empty, PHP associative array.

pub fn with_capacity(size: u32) -> Self

Creates a new, empty, PHP associative array with an initial size.

Parameters

size - The size to initialize the array with.

pub unsafe fn from_ptr(ptr: *mut HashTable, free: bool) -> Result<Self>

Creates a new hash table wrapper. This will not be freed when it goes out of scope in Rust.

Parameters

ptr - The pointer of the actual hash table.

free - Whether the pointer should be freed when the resulting ZendHashTable goes out of scope.

Safety

As a raw pointer is given this function is unsafe, you must ensure that the pointer is valid when calling the function. A simple null check is done but this is not sufficient in most cases.

pub fn len(&self) -> usize

Returns the current number of elements in the array.

pub fn is_empty(&self) -> bool

Returns whether the hash table is empty.

pub fn clear(&mut self)

Clears the hash table, removing all values.

pub fn get<K>(&self, key: K) -> Option<&Zval> where
K: Into<String>,

Attempts to retrieve a value from the hash table with a string key.

Parameters

key - The key to search for in the hash table.

Returns

Some(&Zval) - A reference to the zval at the position in the hash table.

None - No value at the given position was found.

pub fn get_index(&self, key: u64) -> Option<&Zval>

Attempts to retrieve a value from the hash table with an index.

Parameters

key - The key to search for in the hash table.

Returns

Some(&Zval) - A reference to the zval at the position in the hash table.

None - No value at the given position was found.

pub fn remove<K>(&self, key: K) -> Option<()> where
K: Into<String>,

Attempts to remove a value from the hash table with a string key.

Parameters

key - The key to remove from the hash table.

Returns

Some(()) - Key was successfully removed.

None - No key was removed, did not exist.

pub fn remove_index(&self, key: u64) -> Option<()>

Attempts to remove a value from the hash table with a string key.

Parameters

key - The key to remove from the hash table.

Returns

Ok(()) - Key was successfully removed.

None - No key was removed, did not exist.

pub fn insert<K, V>(
&mut self,
key: K,
val: V
) -> Result<HashTableInsertResult<'_>> where
K: Into<String>,
V: IntoZval,

Attempts to insert an item into the hash table, or update if the key already exists. Returns a result containing a HashTableInsertResult, which will indicate a successful insert, with the insert result variants either containing the overwritten value or nothing.

Parameters

key - The key to insert the value at in the hash table.

value - The value to insert into the hash table.

pub fn insert_at_index<V>(
&mut self,
key: u64,
val: V
) -> Result<HashTableInsertResult<'_>> where
V: IntoZval,

Inserts an item into the hash table at a specified index, or updates if the key already exists.

Parameters

key - The index at which the value should be inserted.

val - The value to insert into the hash table.

Returns

Some(&Zval) - The existing value in the hash table that was overriden.

None - The element was inserted.

pub fn push<V>(&mut self, val: V) -> Result<()> where
V: IntoZval,

Pushes an item onto the end of the hash table. Returns a result containing nothing if the element was sucessfully inserted.

Parameters

val - The value to insert into the hash table.

pub fn iter(&self) -> Iter<'_>

Returns an iterator over the hash table.