Struct ext_php_rs::php::types::callable::Callable

pub struct Callable<'a>(_);

Acts as a wrapper around a callable Zval. Allows the owner to call the Zval as if it was a PHP function through the try_call method.

Implementations

impl<'a> Callable<'a>

pub fn new(callable: &'a Zval) -> Result<Self>

Attempts to create a new Callable. Should not need to be used directly, but through the [Zval::callable] method.

Parameters

callable - The underlying Zval that is callable.

Errors

Returns an error if the Zval was not callable.

pub fn try_call(&self, params: Vec<&dyn IntoZval>) -> Result<Zval>

Attempts to call the callable with a list of arguments to pass to the function. Note that a thrown exception inside the callable is not detectable, therefore you should check if the return value is valid rather than unwrapping. Returns a result containing the return value of the function, or an error.

You should not call this function directly, rather through the call_user_func macro.

Parameters

params - A list of parameters to call the function with.