Trait ext_php_rs::php::types::closure::PhpClosure

pub unsafe trait PhpClosure {
    fn invoke(&mut self, ex: &mut ExecutionData, ret: &mut Zval);
}

This is supported on crate feature closure only.

Implemented on types which can be used as PHP closures.

Types must implement the invoke function which will be called when the closure is called from PHP. Arguments must be parsed from the ExecutionData and the return value is returned through the Zval.

This trait is automatically implemented on functions with up to 8 parameters.

Required methods

fn invoke(&mut self, ex: &mut ExecutionData, ret: &mut Zval)

Invokes the closure.