Struct ext_php_rs::php::types::object::ZendClassObject

#[repr(C)]pub struct ZendClassObject<T: Default> { /* fields omitted */ }

A Zend class object which is allocated when a PHP class object is instantiated. Overrides the default handler when the user provides a type T of the struct they want to override with.

Implementations

impl<T: Default> ZendClassObject<T>

pub unsafe fn new_ptr(
ce: *mut ClassEntry,
handlers: *mut ZendObjectHandlers
) -> *mut zend_object

Allocates a new object when an instance of the class is created in the PHP world.

Internal function. The end user functions are generated by the ZendObjectHandler derive macro which generates a function that wraps this function to be exported to C.

Parameters

ce - The class entry that was created.

handlers - A pointer to the object handlers for the class.

Safety

This function is an internal function which is only called from code which is derived using the ZendObjectHandler derive macro. PHP will guarantee that any pointers given to this function will be valid, therefore we can Unwrap them with safety.

pub fn get(ex: &ExecutionData) -> Option<&'static mut Self>

Attempts to retrieve the zend class object container from the zend object contained in the execution data of a function.

Parameters

ex - The execution data of the function.