Trait ext_php_rs::php::types::object::RegisteredClass

pub trait RegisteredClass: Default + Sized where
    Self: 'static, {
    fn get_metadata() -> &'static ClassMetadata<Self>;

    unsafe fn get_property<'a, T: FromZval<'a>>(
        &'a self, 
        name: &str
    ) -> Option<T> { ... }

    unsafe fn set_property(
        &mut self, 
        name: &str, 
        value: impl IntoZval
    ) -> Option<()> { ... }
}

Implemented on Rust types which are exported to PHP. Allows users to get and set PHP properties on the object.

Required methods

fn get_metadata() -> &'static ClassMetadata<Self>

Returns a reference to the class metadata, which stores the class entry and handlers.

This must be statically allocated, and is usually done through the [macro@php_class] macro.

Provided methods

unsafe fn get_property<'a, T: FromZval<'a>>(&'a self, name: &str) -> Option<T>

Attempts to retrieve a property from the class object.

Parameters

name - The name of the property.

Returns

Returns a given type T inside an option which is the value of the zval, or None if the property could not be found.

Safety

Caller must guarantee that the object the function is called on is immediately followed by a zend_object, which is true when the object was instantiated from [ZendClassObject].

unsafe fn set_property(
&mut self,
name: &str,
value: impl IntoZval
) -> Option<()>

Attempts to set the value of a property on the class object.

Parameters

name - The name of the property to set.

value - The value to set the property to.

Returns

Returns nothing in an option if the property was successfully set. Returns none if setting the value failed.

Safety

Caller must guarantee that the object the function is called on is immediately followed by a zend_object, which is true when the object was instantiated from [ZendClassObject].