Trait ext_php_rs::php::types::object::ZendObjectOverride

pub trait ZendObjectOverride {
    extern "C" fn create_object(
        ce: *mut ClassEntry
    ) -> *mut ZendObject;
}

Implemented by the ZendObjectHandler macro on a type T which is used as the T type for ZendClassObject. Implements a function create_object which is passed to a PHP class entry to instantiate the object that will represent an object.

Required methods

extern "C" fn create_object(ce: *mut ClassEntry) -> *mut ZendObject

Creates a new Zend object. Also allocates space for type T on which the trait is implemented on.

Parameters

ce - The class entry that we are creating an object for.