Type Definition ext_php_rs::php::types::object::ZendObject

type ZendObject = zend_object;

Implementations

impl ZendObject

pub fn get_property(&self, name: impl AsRef<str>) -> Result<&Zval>

Attempts to read a property from the Object. Returns a result returning an immutable reference to the Zval if the property exists and can be read, and an Error otherwise.

Parameters

name - The name of the property.

query - The type of query to use when attempting to get a property.

pub fn set_property(
&mut self,
name: impl AsRef<str>,
value: impl Into<Zval>
) -> Result<&Zval>

Attempts to set a property on the object, returning an immutable reference to the Zval if the property can be set.

Parameters

name - The name of the property.

value - The value to set the property to.

pub fn has_property(
&self,
name: impl AsRef<str>,
query: PropertyQuery
) -> Result<bool>

Checks if a property exists on an object. Takes a property name and query parameter, which defines what classifies if a property exists or not. See PropertyQuery for more information.

Parameters

name - The name of the property.

query - The ‘query’ to classify if a property exists.