Trait ext_php_rs::php::types::zval::FromZval

pub trait FromZval<'a>: TryFrom<&'a Zval> {
    const TYPE: DataType;
}

Allows zvals to be converted into Rust types in a fallible way. Reciprocal of the IntoZval trait.

This trait requires the TryFrom trait to be implemented. All this trait does is contain the type of data that is expected when parsing the value, which is used when parsing arguments.

Associated Constants

const TYPE: DataType

The corresponding type of the implemented value in PHP.