Trait ext_php_rs::php::types::zval::IntoZvalDyn

pub trait IntoZvalDyn {
    fn set_zval(&self, zv: &mut Zval, persistent: bool) -> Result<()>;

    fn get_type(&self) -> DataType;

    fn as_zval(&self, persistent: bool) -> Result<Zval> { ... }
}

An object-safe version of the IntoZval trait. This trait is automatically implemented on any type that implements IntoZval, and therefore should not be implemented directly.

Required methods

fn set_zval(&self, zv: &mut Zval, persistent: bool) -> Result<()>

Sets the content of a pre-existing zval. Returns a result containing nothing if setting the content was successful.

Parameters

zv - The Zval to set the content of.

persistent - Whether the contents of the Zval will persist between requests.

fn get_type(&self) -> DataType

Returns the PHP type of the type.

Provided methods

fn as_zval(&self, persistent: bool) -> Result<Zval>

Converts a Rust primitive type into a Zval. Returns a result containing the Zval if successful.

Parameters

persistent - Whether the contents of the Zval will persist between requests.