Struct ext_php_rs::php::args::Arg

pub struct Arg<'a> { /* fields omitted */ }

Represents an argument to a function.

Implementations

impl<'a> Arg<'a>

pub fn new<T: Into<String>>(name: T, _type: DataType) -> Self

Creates a new argument.

Parameters

name - The name of the parameter.

_type - The type of the parameter.

pub fn as_ref(self) -> Self

Sets the argument as a reference.

pub fn allow_null(self) -> Self

Sets the argument as nullable.

pub fn default<T: Into<String>>(self, default: T) -> Self

Sets the default value for the argument.

pub fn val<T: TryFrom<&'a Zval>>(&self) -> Option<T>

Attempts to retrieve the value of the argument. This will be None until the ArgParser is used to parse the arguments.

pub fn zval(&self) -> Option<&'a Zval>

Attempts to return a reference to the arguments internal Zval.

Returns

Some(&Zval) - The internal zval.

None - The argument was empty.

pub fn try_call(&self, params: Vec<&dyn IntoZval>) -> Result<Zval>

Attempts to call the argument as a callable with a list of arguments to pass to the function. Note that a thrown exception inside the callable is not detectable, therefore you should check if the return value is valid rather than unwrapping. Returns a result containing the return value of the function, or an error.

You should not call this function directly, rather through the call_user_func macro.

Parameters

params - A list of parameters to call the function with.