Struct ext_php_rs::php::class::ClassBuilder

pub struct ClassBuilder<'a> { /* fields omitted */ }

Builds a class to be exported as a PHP class.

Implementations

impl<'a> ClassBuilder<'a>

pub fn new<T: Into<String>>(name: T) -> Self

Creates a new class builder, used to build classes to be exported to PHP.

Parameters

name - The name of the class.

pub fn extends(self, parent: &'static ClassEntry) -> Self

Sets the class builder to extend another class.

Parameters

parent - The parent class to extend.

pub fn method(self, func: FunctionEntry, flags: MethodFlags) -> Self

Adds a method to the class.

Parameters

func - The function entry to add to the class.

flags - Flags relating to the function. See MethodFlags.

pub fn property<T: Into<String>>(
self,
name: T,
default: impl IntoZval,
flags: PropertyFlags
) -> Result<Self>

Adds a property to the class. The initial type of the property is given by the type of the given default. Note that the user can change the type.

Returns a result containing the class builder if the property was successfully added.

Parameters

name - The name of the property to add to the class.

default - The default value of the property.

flags - Flags relating to the property. See PropertyFlags.

pub fn constant<T: Into<String>>(
self,
name: T,
value: impl IntoZval
) -> Result<Self>

Adds a constant to the class. The type of the constant is defined by the type of the given default.

Returns a result containing the class builder if the constant was successfully added.

Parameters

name - The name of the constant to add to the class.

value - The value of the constant.

pub fn flags(self, flags: ClassFlags) -> Self

Sets the flags for the class.

Parameters

flags - Flags relating to the class. See ClassFlags.

pub fn object_override<T: ZendObjectOverride>(self) -> Self

Overrides the creation of the Zend object which will represent an instance of this class.

Parameters

T - The type which will override the Zend object. Must implement ZendObjectOverride which can be derived through the ZendObjectHandler derive macro.

pub fn build(self) -> Result<&'static mut ClassEntry>

Builds the class, returning a reference to the class entry.

Errors

Returns an Error variant if the class could not be registered.