Trait ext_php_rs::php::pack::Pack

pub unsafe trait Pack: Clone {
    fn pack_into(vec: Vec<Self>) -> *mut ZendString;

    unsafe fn unpack_into(s: &ZendString) -> Vec<Self>;
}

Used to convert between Zend binary strings and vectors. Useful in conjunction with the pack and unpack functions built-in to PHP.

Safety

The types cannot be ensured between PHP and Rust, as the data is represented as a string when crossing the language boundary. Exercise caution when using these functions.

Required methods

fn pack_into(vec: Vec<Self>) -> *mut ZendString

Packs a given vector into a Zend binary string. Can be passed to PHP and then unpacked using the unpack function. Note you should probably use the set_binary method on the Zval struct instead of this function directly, as there is currently no way to set a ZendString on a Zval directly.

Parameters

vec - The vector to pack into a binary string.

unsafe fn unpack_into(s: &ZendString) -> Vec<Self>

Unpacks a given Zend binary string into a Rust vector. Can be used to pass data from pack in PHP to Rust without encoding into another format. Note that the data must be all one type, as this implementation only unpacks one type.

Safety

This is an unsafe function. There is no way to tell if the data passed from the PHP function is indeed the correct format. Exercise caution when using the unpack functions. In fact, even when used correctly, the results can differ depending on the platform and the size of each type on the platform. Consult the pack function documentation for more details.

Parameters

s - The Zend string containing the binary data.

Loading content...