Type Definition ext_php_rs::php::types::object::ZendObject

type ZendObject = zend_object;

Implementations

impl ZendObject

pub fn get_class_name(&self) -> Result<String>

Attempts to retrieve the class name of the object.

pub fn is_instance<T: RegisteredClass>(&self) -> bool

Checks if the given object is an instance of a registered class with Rust type T.

pub fn get_property(&self, name: &str) -> Result<&Zval>

Attempts to read a property from the Object. Returns a result returning an immutable reference to the Zval if the property exists and can be read, and an Error otherwise.

Parameters

name - The name of the property.

query - The type of query to use when attempting to get a property.

pub fn set_property(
&mut self,
name: &str,
value: impl IntoZval
) -> Result<&Zval>

Attempts to set a property on the object, returning an immutable reference to the Zval if the property can be set.

Parameters

name - The name of the property.

value - The value to set the property to.

pub fn has_property(&self, name: &str, query: PropertyQuery) -> Result<bool>

Checks if a property exists on an object. Takes a property name and query parameter, which defines what classifies if a property exists or not. See PropertyQuery for more information.

Parameters

name - The name of the property.

query - The ‘query’ to classify if a property exists.

pub fn get_properties(&self) -> Result<ZendHashTable<'_>>

Attempts to retrieve the properties of the object. Returned inside a Zend Hashtable.