Trait ext_php_rs::php::types::zval::IntoZvalDyn

pub trait IntoZvalDyn {
    fn as_zval(&self, persistent: bool) -> Result<Zval>;

    fn get_type(&self) -> DataType;
}

An object-safe version of the IntoZval trait.

This trait is automatically implemented on any type that implements both IntoZval and Clone. You avoid implementing this trait directly, rather implement these two other traits.

Required methods

fn as_zval(&self, persistent: bool) -> Result<Zval>

Converts a Rust primitive type into a Zval. Returns a result containing the Zval if successful. self is cloned before being converted into a zval.

Parameters

persistent - Whether the contents of the Zval will persist between requests.

fn get_type(&self) -> DataType

Returns the PHP type of the type.