Type Definition ext_php_rs::php::types::zval::Zval

type Zval = zval;

Zend value. Represents most data types that are in the Zend engine.

Implementations

impl<'a> Zval

pub fn long(&self) -> Option<ZendLong>

Returns the value of the zval if it is a long.

pub fn bool(&self) -> Option<bool>

Returns the value of the zval if it is a bool.

pub fn double(&self) -> Option<f64>

Returns the value of the zval if it is a double.

pub fn string(&self) -> Option<String>

Returns the value of the zval if it is a string.

If the zval does not contain a string, the function will check if it contains a double or a long, and if so it will convert the value to a String and return it.

pub fn str(&'a self) -> Option<&'a str>

Returns the value of the zval if it is a string.

Note that this functions output will not be the same as [Zval::string], as this function does not attempt to convert other types into a String, as it could not pass back a &str in those cases.

pub fn binary<T: Pack>(&self) -> Option<Vec<T>>

Returns the value of the zval if it is a string and can be unpacked into a vector of a given type. Similar to the unpack in PHP, except you can only unpack one type.

Safety

There is no way to tell if the data stored in the string is actually of the given type. The results of this function can also differ from platform-to-platform due to the different representation of some types on different platforms. Consult the pack function documentation for more details.

pub fn resource(&self) -> Option<*mut zend_resource>

Returns the value of the zval if it is a resource.

pub fn array(&self) -> Option<ZendHashTable<'a>>

Returns the value of the zval if it is an array.

pub fn object(&self) -> Option<&mut ZendObject>

Returns the value of the zval if it is an object.

pub fn reference(&self) -> Option<&mut Zval>

Returns the value of the zval if it is a reference.

pub fn callable(&'a self) -> Option<Callable<'a>>

Returns the value of the zval if it is callable.

pub fn try_call(&self, params: Vec<&dyn IntoZval>) -> Result<Zval>

Attempts to call the zval as a callable with a list of arguments to pass to the function. Note that a thrown exception inside the callable is not detectable, therefore you should check if the return value is valid rather than unwrapping. Returns a result containing the return value of the function, or an error.

You should not call this function directly, rather through the call_user_func macro.

Parameters

params - A list of parameters to call the function with.

pub fn get_type(&self) -> Result<DataType>

Returns the type of the Zval.

pub fn is_long(&self) -> bool

Returns true if the zval is a long, false otherwise.

pub fn is_null(&self) -> bool

Returns true if the zval is null, false otherwise.

pub fn is_true(&self) -> bool

Returns true if the zval is true, false otherwise.

pub fn is_false(&self) -> bool

Returns true if the zval is false, false otherwise.

pub fn is_bool(&self) -> bool

Returns true if the zval is a bool, false otherwise.

pub fn is_double(&self) -> bool

Returns true if the zval is a double, false otherwise.

pub fn is_string(&self) -> bool

Returns true if the zval is a string, false otherwise.

pub fn is_resource(&self) -> bool

Returns true if the zval is a resource, false otherwise.

pub fn is_array(&self) -> bool

Returns true if the zval is an array, false otherwise.

pub fn is_object(&self) -> bool

Returns true if the zval is an object, false otherwise.

pub fn is_reference(&self) -> bool

Returns true if the zval is a reference, false otherwise.

pub fn is_callable(&self) -> bool

Returns true if the zval is callable, false otherwise.

pub fn set_string(&mut self, val: &str, persistent: bool) -> Result<()>

Sets the value of the zval as a string. Returns nothing in a result when successful.

Parameters

val - The value to set the zval as.

persistent - Whether the string should persist between requests.

pub fn set_binary<T: Pack, U: AsRef<[T]>>(&mut self, val: U)

Sets the value of the zval as a binary string.

Parameters

val - The value to set the zval as.

pub fn set_interned_string(&mut self, val: &str) -> Result<()>

Sets the value of the zval as a interned string. Returns nothing in a result when successful.

Parameters

val - The value to set the zval as.

pub fn set_long<T: Into<ZendLong>>(&mut self, val: T)

Sets the value of the zval as a long.

Parameters

val - The value to set the zval as.

pub fn set_double<T: Into<c_double>>(&mut self, val: T)

Sets the value of the zval as a double.

Parameters

val - The value to set the zval as.

pub fn set_bool<T: Into<bool>>(&mut self, val: T)

Sets the value of the zval as a boolean.

Parameters

val - The value to set the zval as.

pub fn set_null(&mut self)

Sets the value of the zval as null.

This is the default of a zval.

pub fn set_resource(&mut self, val: *mut zend_resource)

Sets the value of the zval as a resource.

Parameters

val - The value to set the zval as.

pub fn set_object(&mut self, val: &ZendObject, _copy: bool)

Sets the value of the zval as an object.

Parameters

val - The value to set the zval as.

copy - Whether to copy the object or pass as a reference.

pub fn set_array(&mut self, val: ZendHashTable<'_>)

Sets the value of the zval as an array. Returns nothng in a result on success.

Parameters

val - The value to set the zval as.